+++ python/apsis/schedule/stop.py
